requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.2.0
python-dotenv==0.20.0
Werkzeug==2.2.2
zipp==3.8.1
